fixed types->validate-request circular dep
