+++ src/lib/adjudicators.test.ts
@@ -4,8 +4,10 @@
 import { expect, describe, it } from "@jest/globals";
 import * as sut from "./adjudicators";
 import { KubernetesObject } from "kubernetes-fluent-client";
-import { AdmissionRequest, Binding, Event } from "./types";
-import { DeepPartial, Operation } from "./mutate-types";
+import { Binding, Event } from "./types";
+import { Operation } from "./mutate-types";
+import { AdmissionRequest } from "./validate-types";
+import { DeepPartial } from "./mutate-types";
 
 describe("definesDeletionTimestamp", () => {
   //[ Binding, result ]
