+++ src/lib/filter.ts
@@ -1,7 +1,8 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: 2023-Present The Pepr Authors
 
-import { AdmissionRequest, Binding } from "./types";
+import { Binding } from "./types";
+import { AdmissionRequest } from "./validate-types";
 import { Operation } from "./mutate-types";
 import {
   carriesIgnoredNamespace,
