+++ src/lib/module.ts
@@ -5,7 +5,8 @@ import { Capability } from "./capability";
 import { Controller } from "./controller";
 import { ValidateError } from "./errors";
 import { MutateResponse, ValidateResponse, WebhookIgnore } from "./k8s";
-import { CapabilityExport, AdmissionRequest } from "./types";
+import { CapabilityExport } from "./types";
+import { AdmissionRequest } from "./validate-types";
 import { setupWatch } from "./watch-processor";
 import { Log } from "../lib";
