+++ src/lib/mutate-request.ts
@@ -1,11 +1,12 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: 2023-Present The Pepr Authors
 
-import { Operation, AdmissionRequest, DeepPartial } from "./mutate-types";
+import { AdmissionRequest, DeepPartial } from "./mutate-types";
 import { KubernetesObject } from "kubernetes-fluent-client";
 import { clone, mergeDeepRight } from "ramda";
 import { Logger } from "pino";
 import { GenericClass } from "kubernetes-fluent-client";
+import { Operation } from "./mutate-types";
 
 // MutateAction type for handling mutation callbacks
 export type MutateAction<T extends GenericClass, K extends KubernetesObject = InstanceType<T>> = (
